Project_final v.2.0 after first review

+++ yatube/posts/views.py
@@ -1,87 +1,19 @@
-from django.core.cache import cache
 from django.shortcuts import render, get_object_or_404, redirect
 from .models import Post, Group, User, Follow
-
-from django.core.paginator import Paginator, Page
+from .cached_paginator import create_page, create_page_not_cached
 from posts.forms import PostForm, CommentForm
 from django.contrib.auth.decorators import login_required
 
 POSTS_ON_PAGE: int = 10
-CACHE_TIMEOUT: int = 20
-
-
-class CachedPaginator(Paginator):
-    """A paginator that caches the results on a page by page basis."""
-    def __init__(self,
-                 object_list,
-                 per_page,
-                 cache_key,
-                 cache_timeout=300,
-                 orphans=0,
-                 allow_empty_first_page=True):
-        super(CachedPaginator, self).__init__(object_list,
-                                              per_page,
-                                              orphans,
-                                              allow_empty_first_page)
-        self.cache_key = cache_key
-        self.cache_timeout = cache_timeout
-
-    def page(self, number):
-        """
-        Returns a Page object for the given 1-based page number.
-
-        This will attempt to pull the results out of the cache first, based on
-        the number of objects per page and the requested page number. If not
-        found in the cache, it will pull a fresh list and then cache that
-        result.
-        """
-        if number is None:
-            number = 1
-        number = self.validate_number(number)
-        cached_object_list = cache.get(self.build_cache_key(number), None)
-
-        if cached_object_list is not None:
-            page = Page(cached_object_list, number, self)
-        else:
-            page = super(CachedPaginator, self).page(number)
-            # Since the results are fresh, cache it.
-            cache.set(self.build_cache_key(number),
-                      page.object_list,
-                      self.cache_timeout)
-
-        return page
-
-    def build_cache_key(self, page_number):
-        """Appends the relevant pagination bits to the cache key."""
-        return "%s:%s:%s" % (self.cache_key, self.per_page, page_number)
-
-
-def create_page(posts: Post,
-                page_number: int,
-                posts_on_page: int,
-                cache_key: str) -> Page:
-    paginator = CachedPaginator(posts,
-                                posts_on_page,
-                                cache_key,
-                                CACHE_TIMEOUT
-                                )
-    page_obj = paginator.page(page_number)
-    return page_obj
-
-
-def create_page_not_cached(posts: Post,
-                           page_number: int,
-                           posts_on_page: int) -> Page:
-    paginator = Paginator(posts, posts_on_page)
-    page_obj = paginator.get_page(page_number)
-    return page_obj
 
 
 def index(request):
     template = 'posts/index.html'
     posts = Post.objects.select_related('group').all()
-    page_number = request.GET.get('page')
-    page_obj = create_page(posts, page_number, POSTS_ON_PAGE, 'index_page')
+    page_obj = create_page(posts,
+                           request.GET.get('page'),
+                           POSTS_ON_PAGE,
+                           'index_page')
     context = {
         'page_obj': page_obj,
     }
@@ -92,9 +24,8 @@ def group_posts(request, slug):
     template = 'posts/group_list.html'
     group = get_object_or_404(Group, slug=slug)
     posts = group.posts.all()
-    page_number = request.GET.get('page')
     page_obj = create_page(posts,
-                           page_number,
+                           request.GET.get('page'),
                            POSTS_ON_PAGE,
                            f'group_page_{slug}')
     context = {
@@ -109,16 +40,14 @@ def profile(request, username):
     user = get_object_or_404(User, username=username)
     is_author = (user == request.user)
     posts = Post.objects.filter(author=user)
-    page_number = request.GET.get('page')
+    following = False
     page_obj = create_page(posts,
-                           page_number,
+                           request.GET.get('page'),
                            POSTS_ON_PAGE,
                            f'profile_page_{username}')
     if request.user.is_authenticated:
         following = Follow.objects.filter(user=request.user,
                                           author=user).exists()
-    else:
-        following = None
     context = {
         'username': user,
         'page_obj': page_obj,
@@ -131,11 +60,9 @@ def profile(request, username):
 def post_detail(request, post_id: int):
     template = 'posts/post_detail.html'
     post = Post.objects.get(pk=post_id)
-    author_posts = post.author.posts.all()
     comments = post.comments.all()
     context = {
         'post': post,
-        'author_posts': author_posts,
         'form': CommentForm(),
         'comments': comments,
     }
@@ -144,12 +71,10 @@ def post_detail(request, post_id: int):
 
 @login_required
 def post_create(request):
-    groups = Group.objects.all()
     form = PostForm(request.POST or None,
                     files=request.FILES or None)
     context = {
         'form': form,
-        'groups': groups,
     }
     if form.is_valid():
         form = form.save(commit=False)
@@ -163,7 +88,6 @@ def post_create(request):
 
 @login_required
 def post_edit(request, post_id: int):
-    groups = Group.objects.all()
     post = Post.objects.get(pk=post_id)
     if request.user != post.author:
         return redirect('posts:post_detail', post_id)
@@ -172,7 +96,6 @@ def post_edit(request, post_id: int):
                     instance=post)
     context = {
         'form': form,
-        'groups': groups,
         'post': post,
         'is_edit': True,
     }
@@ -200,8 +123,9 @@ def add_comment(request, post_id):
 def follow_index(request):
     user = request.user
     posts = Post.objects.filter(author__following__user=user)
-    page_number = request.GET.get('page')
-    page_obj = create_page_not_cached(posts, page_number, POSTS_ON_PAGE)
+    page_obj = create_page_not_cached(posts,
+                                      request.GET.get('page'),
+                                      POSTS_ON_PAGE)
     context = {'page_obj': page_obj, }
     return render(request, 'posts/index.html', context)
 
@@ -220,5 +144,7 @@ def profile_follow(request, username):
 def profile_unfollow(request, username):
     author = get_object_or_404(User, username=username)
     user = request.user
-    Follow.objects.filter(user=user, author=author).delete()
+    if Follow.objects.filter(user=user,
+                             author=author).exists():
+        Follow.objects.filter(user=user, author=author).delete()
     return redirect('posts:profile', username=username)
